Moved model functionallity from ReigsterActivity to new RegisterPresenter
