+++ application/app/src/main/java/com/yahyeet/boardbook/activity/RegisterActivity.java
@@ -12,9 +12,14 @@ import android.widget.Button;
 import android.widget.EditText;
 
 import com.yahyeet.boardbook.R;
+import com.yahyeet.boardbook.presenter.BoardbookSingleton;
+import com.yahyeet.boardbook.presenter.IRegisterPresenter;
+import com.yahyeet.boardbook.presenter.RegisterPresenter;
 
 public class RegisterActivity extends AppCompatActivity {
 
+
+    IRegisterPresenter registerPresenter;
 
     EditText emailInput;
     EditText passwordInput;
@@ -30,6 +35,10 @@ public class RegisterActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_register);
+
+        // TODO: Make accountManagerActivity (To be created) assign loginPresenter and registerPresenter
+        registerPresenter = new RegisterPresenter(this);
+
 
         emailInput = findViewById(R.id.emailRegisterInput);
         passwordInput = findViewById(R.id.passwordRegisterInput);
@@ -108,52 +117,12 @@ public class RegisterActivity extends AppCompatActivity {
         }
     }
 
-
-    /*
-    /**
-     * Method reads info in email and password fields
-     * @return a String[] in the form of {email, password}.
-
-    public String[] FetchRegisterFields(){
-
-
-        // TODO Send information in a better way :S
-        return  new String[]{emailInput.getText().toString(), passwordInput.getText().toString(), userInput.getText().toString()};
-
-    }
-    */
-
     /**
      *  Method makes a new account if the "Make a New Account" button has been tapped
      * @param view is the visual object (ex a button) the method is bound to
      */
     public void RegisterAccount(View view){
-
-
-        ProgressDialog progress = new ProgressDialog(this);
-        // TODO Make better loading text
-        progress.setTitle("Loading");
-        progress.setMessage("Wait while loading...");
-        progress.setCancelable(false); // disable dismiss by tapping outside of the dialog
-        progress.show();
-
-
-        BoardbookSingleton.getInstance().getAuthHandler().signup(
-                emailInput.getText().toString(), passwordInput.getText().toString(), userInput.getText().toString()).thenAccept(u -> {
-            // access logged in user from "u"
-
-            progress.dismiss();
-            finish();
-
-        }).exceptionally(e -> {
-            // Handle error ("e")
-
-            progress.dismiss();
-            e.printStackTrace();
-
-            return null;
-        });
-
+        registerPresenter.signup(emailInput.getText().toString(), passwordInput.getText().toString(), userInput.getText().toString());
     }
