+++ application/app/src/main/java/com/yahyeet/boardbook/presenter/ILoginPresenter.java
@@ -6,8 +6,4 @@ public interface ILoginPresenter {
 
     void login(String email, String password);
 
-    LoginActivity getLoginActivity();
-
-    void setLoginActivity(LoginActivity loginActivity);
-
 }
