+++ application/app/src/main/java/com/yahyeet/boardbook/presenter/IRegisterPresenter.java
@@ -1,0 +1,8 @@
+package com.yahyeet.boardbook.presenter;
+
+import com.yahyeet.boardbook.activity.RegisterActivity;
+
+public interface IRegisterPresenter {
+
+    void signup(String email, String password, String username);
+}
